// src/Constants/constants.d.ts
import { EInputTypes } from "../types/globals";
export declare const INPUT_TYPES: {
    label: string;
    value: EInputTypes;
}[];
export declare const INPUT_PROPS_MAP: {
    text: {
        props: string[];
        validations: string[];
        additionalProps: string[];
    };
    select: {
        props: string[];
        validations: string[];
        additionalProps: string[];
    };
    number: {
        props: string[];
        validations: string[];
        additionalProps: string[];
    };
};
export declare const PROPS: string[];
export declare const VALIDATIONS: string[];
export declare const OTHER_PROPS: string[];